requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
pytest==5.3.4
flake8==3.7.9
djangorestframework==3.11.0
ipdb